Added mandatory/optional choice to validators
If a validator is set as `mandatory', then the validation process has to
be run before allowing to click submit button when adding an entity to a
federation. On the other hand, `optional' validator (i.e. mandatory
checkbox is not checked) just produces a warning when the validation
process fail.

More info to come when finished :)