Colour validator button depending on results
Before a join federation request is accepted or rejected, it's possible
to validate it first. When a validator fails the button is coloured red.
When a validator passes the button is coloured green. Just eye-candy,
but might help when more validators are used in a federation.

P.S. It's just a quick hack, maybe you'll find a better way to implement
it in JS. Don't know jQuery too well.