Mandatory validator check when accepting an entity
All mandatory validators have to be run and they have pass successfully
before a new entity is allowed to be registered. If not all mandatory
validators are run or finish correctly, it's not possible to accept such
an entity. (Validator can be set as mandatory in federation settings
using a checkbox.)

Also, colouring buttons to reflect validation status (passed -> green,
failed -> red) is rewritten into jQuery.